requirements.txt
Orange3>=3.21.0
serverfiles
requests
requests-cache
genesis-pyapi
scipy
numpy
shapely
